# generate-boundaries perf improvements